requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1;python_version<"3.9"
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.3
django-celery-beat==2.5.0
django-taggit==4.0.0
django-timezone-field==5.1
djangorestframework==3.14.0
drf-spectacular==0.26.3
flower==2.0.0
humanize==4.7.0
idna==3.4
importlib-resources==6.0.0
inflection==0.5.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kombu==5.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
pkgutil-resolve-name==1.3.10
prometheus-client==0.17.0
prompt-toolkit==3.0.39
Pygments==2.15.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
referencing==0.29.1
requests==2.31.0
rich==13.4.2
rpds-py==0.8.10
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
zipp==3.16.0